inngest-example/pyproject.toml
[project]
name = "inngest-example"
version = "0.1.0"
description = "Inngest Durable Execution Evaluation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "inngest>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "openai>=1.60.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
